API

VMRobotControl.AbstractJointData

VMRobotControl.CompiledCoordID

VMRobotControl.CompiledMechanism

VMRobotControl.CompiledVirtualMechanismSystem

VMRobotControl.ComponentData

VMRobotControl.ConstCoord

VMRobotControl.CoordDifference

VMRobotControl.CoordSlice

VMRobotControl.CoordStack

VMRobotControl.CoordSum

VMRobotControl.CoordinateData

VMRobotControl.DiodicDamper

VMRobotControl.Dissipation

VMRobotControl.EltypeCache

VMRobotControl.ForceSource

VMRobotControl.FrameAngularVelocity

VMRobotControl.FrameOrigin

VMRobotControl.FramePoint

VMRobotControl.GaussianSpring

VMRobotControl.GenericComponent

VMRobotControl.GravityCompensator

VMRobotControl.Inertance

VMRobotControl.Inertia

VMRobotControl.JointSubspace

VMRobotControl.LinearDamper

VMRobotControl.LinearInerter

VMRobotControl.LinearSpring

VMRobotControl.Mechanism

VMRobotControl.MechanismJoint

VMRobotControl.PointMass

VMRobotControl.PowerSource

VMRobotControl.QuaternionAttitude

VMRobotControl.RBTree

VMRobotControl.RSON.RSONParserConfig

VMRobotControl.RectifiedDamper

VMRobotControl.ReferenceCoord

VMRobotControl.ReferenceJoint

VMRobotControl.Rigid

VMRobotControl.RotatedCoord

VMRobotControl.Storage

VMRobotControl.TanhDamper

VMRobotControl.TanhSpring

VMRobotControl.TimeFuncJoint

VMRobotControl.Transforms.EulerVector

VMRobotControl.Transforms.SpatialAcceleration

VMRobotControl.Transforms.Transform

VMRobotControl.Transforms.Twist

VMRobotControl.TypeStableCollections.TypeStableCollection

VMRobotControl.TypeStableCollections.TypeStableCollection

VMRobotControl.TypeStableCollections.TypeStableIdx

VMRobotControl.URDF.URDFParserConfig

VMRobotControl.UnrotatedCoord

VMRobotControl.VirtualMechanismSystem

VMRobotControl.Prismatic

VMRobotControl.RSON.parseRSON

VMRobotControl.RSON.parseRSONString

VMRobotControl.RSON.parseRSONVersion

VMRobotControl.RSON.serializeRSON

VMRobotControl.Rail

VMRobotControl.Revolute

VMRobotControl.Spherical

VMRobotControl.Splines.curve_derivative

VMRobotControl.Splines.curve_position

VMRobotControl.Splines.curve_second_derivative

VMRobotControl.Transforms.AxisAngleDerivative

VMRobotControl.Transforms.angular_velocity_prematrix

VMRobotControl.Transforms.angular_velocity_prematrix_derivative

VMRobotControl.Transforms.quaternion_derivative_propagation

VMRobotControl.Transforms.quaternion_derivative_propagation_derivative

VMRobotControl.Transforms.quatmul_geodual_bivector_matrix

VMRobotControl.Transforms.skew

VMRobotControl.TypeStableCollections.extend

VMRobotControl.TypeStableCollections.push_return_index!

VMRobotControl.URDF.parseURDF

VMRobotControl.URDF.parseURDFString

VMRobotControl._add_jacobian_transpose_times_force!

VMRobotControl._construct_coordinate_graph

VMRobotControl.acceleration_kinematics!

VMRobotControl.add_component!

VMRobotControl.add_coordinate!

VMRobotControl.add_frame!

VMRobotControl.add_gravity_compensation!

VMRobotControl.add_gravity_compensator!

VMRobotControl.add_inertia!

VMRobotControl.add_joint!

VMRobotControl.assemble_state

VMRobotControl.autodiff_jacobian_column

VMRobotControl.child_frame_twist

VMRobotControl.child_rbstate

VMRobotControl.compile

VMRobotControl.components

VMRobotControl.config_size

VMRobotControl.configuration

VMRobotControl.control_step!

VMRobotControl.coordinates

VMRobotControl.dissipations

VMRobotControl.dynamics!

VMRobotControl.frame_ancestor_matrix

VMRobotControl.frames

VMRobotControl.generalized_force!

VMRobotControl.generic_components

VMRobotControl.get_angular_acceleration

VMRobotControl.get_angular_jacobian

VMRobotControl.get_angular_vel

VMRobotControl.get_angular_vpa

VMRobotControl.get_compiled_componentID

VMRobotControl.get_compiled_coordID

VMRobotControl.get_compiled_frameID

VMRobotControl.get_compiled_jointID

VMRobotControl.get_generalized_force

VMRobotControl.get_inertance_matrix

VMRobotControl.get_linear_acceleration

VMRobotControl.get_linear_jacobian

VMRobotControl.get_linear_vel

VMRobotControl.get_linear_vpa

VMRobotControl.get_ode_dynamics

VMRobotControl.get_ode_problem

VMRobotControl.get_q

VMRobotControl.get_q̇

VMRobotControl.get_q̈

VMRobotControl.get_rbstate

VMRobotControl.get_t

VMRobotControl.get_transform

VMRobotControl.get_u

VMRobotControl.inertance_matrix!

VMRobotControl.inertances

VMRobotControl.jacobian

VMRobotControl.jacobian_column

VMRobotControl.jacobians!

VMRobotControl.joint_relative_acceleration

VMRobotControl.joint_relative_transform

VMRobotControl.joint_relative_twist

VMRobotControl.joint_relative_vpa

VMRobotControl.joints

VMRobotControl.kinematics!

VMRobotControl.name

VMRobotControl.ndof

VMRobotControl.new_control_cache

VMRobotControl.new_dynamics_cache

VMRobotControl.new_jacobians_cache

VMRobotControl.new_kinematics_cache

VMRobotControl.new_rbstates_cache

VMRobotControl.num_frames

VMRobotControl.opspace_force

VMRobotControl.precompute!

VMRobotControl.reassign_coords

VMRobotControl.reassign_frames

VMRobotControl.reassign_joints

VMRobotControl.remake

VMRobotControl.rigid_joint_twist

VMRobotControl.root_frame

VMRobotControl.state_idxs

VMRobotControl.storages

VMRobotControl.stored_energy

VMRobotControl.velocity

VMRobotControl.velocity_kinematics!

VMRobotControl.velocity_size

VMRobotControl.visuals

VMRobotControl.zero_q

VMRobotControl.zero_q̇

VMRobotControl.zero_q̈

VMRobotControl.zero_u

Loading from RSON/URDF

VMRobotControl.RSON.parseRSON — Function

parseRSON(filepath::AbstractString, [cfg::RSONParserConfig])

Parse a Mechanism or VirtualMechanismSystem from an RSON file (with a .rson extension).

See also parseRSONString, serializeRSON, RSONParserConfig.

VMRobotControl.URDF.parseURDF — Function

parseURDF(filepath::AbstractString, [cfg::URDFParserConfig])
parseURDF(cfg::URDFParserConfig)

Parse a Mechanism from a URDF file. If filepath is provided, a default config is used. If ONLY cfg is provided, the URDF file path must be set in the config.

See also parseURDFString, [URDFParserConfig]

VMRobotControl.RSON.RSONParserConfig — Type

mutable struct RSONParserConfig(; kwargs...)

Configuration for the RSON parser.

Fields

Options

element_type::Type: Type of the elements in the mechanism. Default is Float64.

parse_visuals::Bool: Whether to parse visuals. Default is true.

error_on_not_supported::Bool: Whether to throw an error if an unsupported feature is encountered. Default is false.

error_on_not_recognized::Bool: Whether to throw an error if an unrecognized key is encountered. Default is true.

suppress_warnings::Bool: Whether to suppress warnings. Default is false.

default_material::RGBA: Default material for visuals. Default is RGBA(0.8, 0.8, 0.8, 1.0).

Fields used by the parser

rson_path::String: Path to the RSON file parsed. (Used to locate meshes)

rson_version::Tuple{Int, Int, Int}: Version of the RSON file parsed.

named_materials::Dict{String, RGBA}: Named materials for visuals. Default is an empty dictionary.

Internal

name_stack::Vector{String}: Stack of names for error messages.

VMRobotControl.URDF.URDFParserConfig — Type

mutable struct URDFParserConfig

Used to configure the URDF parser. WARNING: may be modified during parsing (e.g. materials are added to named materials).

Fields

urdf_path::String: Path to the URDF file.

element_type::Type: eltype to build the mechanism with, default Float64.

parse_visuals::Bool: Whether to parse visual elements, default true.

error_on_not_supported::Bool: Whether to error on unsupported URDF features, default false.

error_on_not_recognized::Bool: Whether to error on unrecognized URDF features, default true.

suppress_warnings::Bool: Whether to suppress warnings, default false.

default_material::RGBA: Default colour to use for visuals, default RGBA(0.8, 0.8, 0.8, 1.0).

named_materials::Dict{String, RGBA}: Named materials to use for visuals, default Dict().

VMRobotControl.RSON.serializeRSON — Function

serializeRSON(filepath, m::Union{Mechanism, VirtualMechanismSystem})

Takes a mechanism or virtual mechanism system and serializes it to a file at filepath in the RSON format.

When handling visuals, the mesh data is serialized as a list of vertices and faces.

Does not guarantee that a parsed RSON is serialized back into an identical RSON, but the mechanism should be the same.

Building

Types

VMRobotControl.Mechanism — Type

Mechanism{T}(name; with_root_frame=true)

A rigid body 'Mechanism', composed of reference frames, joints, coordinates and dynamic components.

VMRobotControl.VirtualMechanismSystem — Type

VirtualMechanismSystem{T}(name)
VirtualMechanismSystem(name, robot::Mechanism{T}, [virtual_mechanism::Mechanism{T}])

Construct a virtual mechanism system. If only a name is provided, then an empty virtual mechanism system is created, and the robot/virtual mechanism can be accessed by the robot and virtual_mechanism fields.

Coordinates and components can then be added to the virtual mechanism system, and these can act inbetween the robot and the virtual mechanism. Coordinates on the robot/virtual mechanism can be referenced using the naming syntax ".robot.coordID" or ".virtual_mechanism.coordID".

Note: the robot and virtual mechanism are not copied, so modifying them after creating the virtual mechanism system WILL modify the virtual mechanism system.

See also Mechanism, add_coordinate!, add_component!

Accessing Mechanisms/Virtual Mechanism Systems

VMRobotControl.name — Function

name(m::Union{Mechanism, CompiledMechanism, VirtualMechanismSystem, CompiledVirtualMechanismSystem})

Returns the name of the mechanism.

VMRobotControl.frames — Function

frames(m::Union{Mechanism, CompiledMechanism})

Returns a vector of the names of the frames in the mechanism.

VMRobotControl.num_frames — Function

num_frames(m::Union{Mechanism, CompiledMechanism})

Returns the number of reference frames in the mechanism.

VMRobotControl.root_frame — Function

root_frame(m::Union{Mechanism, CompiledMechanism})

Returns the name of the first frame of the mechanism. This should be the root frame of the tree and should be called "root_frame".

VMRobotControl.joints — Function

joints(m::Union{Mechanism, CompiledMechanism})

Returns joints in the mechanism.

If m<:Union{Mechanism, VirtualMechanismSystem} this is an OrderedDict. Mutating it WILL mutate the mechanism. If m<:Union{CompiledMechanism, CompiledVirtualMechanismSystem} this is a TypeStableCollection. Mutating the returned collection WILL mutate the mechanism.

It is recommended to use add_joint! to add joints to the mechanism, and get_compiled_jointID to get an ID/index for modifying joints in compiled systems.

VMRobotControl.coordinates — Function

coordinates(m::Union{Mechanism, CompiledMechanism, VirtualMechanismSystem, CompiledVirtualMechanismSystem})

Returns the coordinates of the system.

If m<:Union{Mechanism, VirtualMechanismSystem} this is a Dict. Mutating the returned Dict WILL mutate the mechanism. If m<:Union{CompiledMechanism, CompiledVirtualMechanismSystem} this is a TypeStableCollection. Mutating the returned collection WILL mutate the mechanism.

It is recommended to use add_coordinate! to add coordinates to the mechanism, and get_compiled_coordID to get an ID/index for modifying coordinates in compiled systems.

VMRobotControl.components — Function

components(m::Union{Mechanism, CompiledMechanism, VirtualMechanismSystem, CompiledVirtualMechanismSystem})

Returns the dynamic components of the mechanism (e.g. springs/dampers/masses).

If m<:Union{Mechanism, VirtualMechanismSystem} this is a Dict. Mutating the returned Dict WILL mutate the mechanism. If m<:Union{CompiledMechanism, CompiledVirtualMechanismSystem} this is a TypeStableCollection. Mutating the returned collection WILL mutate the mechanism.

It is recommended to use add_component! to add components to the mechanism, and get_compiled_componentID to get an ID/index for modifying components in compiled systems.

VMRobotControl.config_size — Function

config_size(m::Union{Mechanism, CompiledMechanism})

Returns the number of configuration variables in the mechanism (e.g. the number of joint angles), the length of the joint configuration vector q.

VMRobotControl.velocity_size — Function

velocity_size(m::Union{Mechanism, CompiledMechanism})

Returns the number of velocity variables in the mechanism (e.g. the number of joint velocities), the length of the joint velocity vector q̇.

VMRobotControl.ndof — Function

ndof(m::Union{Mechanism, CompiledMechanism, VirtualMechanismSystem, CompiledVirtualMechanismSystem})

The number of degrees of freedom of the system.

VMRobotControl.inertances — Function

inertances(m::Union{Mechanism, CompiledMechanism, VirtualMechanismSystem, CompiledVirtualMechanismSystem})

Returns the inertance components of the system.

If m<:Union{Mechanism, VirtualMechanismSystem} this is a Dict formed by filtering the components by Inertance type: mutating the returned Dict WILL NOT mutate the mechanism.

If m<:Union{CompiledMechanism, CompiledVirtualMechanismSystem} this is a TypeStableCollection made from a subset of the vectors in the components TypeStableCollection. Mutating the returned collection WILL mutate the mechanism.

It is recommended to use add_component! to add components to the mechanism, and get_compiled_componentID to get an ID/index for modifying components in compiled systems.

VMRobotControl.storages — Function

storages(m::AbstractMechanism)

Returns the storage components from mechanism.

If m<:Union{Mechanism, VirtualMechanismSystem} this is a Dict formed by filtering the components by Storage type: mutating the returned Dict WILL NOT mutate the mechanism.

If m<:Union{CompiledMechanism, CompiledVirtualMechanismSystem} this is a TypeStableCollection made from a subset of the vectors in the components TypeStableCollection. Mutating the returned collection WILL mutate the mechanism.

It is recommended to use add_component! to add components to the mechanism, and get_compiled_componentID to get an ID/index for modifying components in compiled systems.

VMRobotControl.dissipations — Function

dissipations(m::AbstractMechanism)

Returns the dissipation components from mechanism.

If m<:Union{Mechanism, VirtualMechanismSystem} this is a Dict formed by filtering the components by Dissipation type: mutating the returned Dict WILL NOT mutate the mechanism.

If m<:Union{CompiledMechanism, CompiledVirtualMechanismSystem} this is a TypeStableCollection made from a subset of the vectors in the components TypeStableCollection. Mutating the returned collection WILL mutate the mechanism.

It is recommended to use add_component! to add components to the mechanism, and get_compiled_componentID to get an ID/index for modifying components in compiled systems.

VMRobotControl.generic_components — Function

generic_components(m::AbstractMechanism)

Returns the generic components from mechanism.

If m<:Union{Mechanism, VirtualMechanismSystem} this is a Dict formed by filtering the components by GenericComponent type: mutating the returned Dict WILL NOT mutate the mechanism.

If m<:Union{CompiledMechanism, CompiledVirtualMechanismSystem} this is a TypeStableCollection made from a subset of the vectors in the components TypeStableCollection. Mutating the returned collection WILL mutate the mechanism.

VMRobotControl.visuals — Function

visuals(m::AbstractMechanism)

Returns the visual components from mechanism.

If m<:Union{Mechanism, VirtualMechanismSystem} this is a Dict formed by filtering the components by Visual type: mutating the returned Dict WILL NOT mutate the mechanism.

If m<:Union{CompiledMechanism, CompiledVirtualMechanismSystem} this is a TypeStableCollection made from a subset of the vectors in the components TypeStableCollection. Mutating the returned collection WILL mutate the mechanism.

It is recommended to use add_component! to add components to the mechanism, and get_compiled_componentID to get an ID/index for modifying components in compiled systems.

Building methods

VMRobotControl.add_frame! — Function

add_frame!(m::Mechanism; id::String)

Add a frame to a mechanism. The name must be unique.

See also frames, add_joint!

VMRobotControl.add_joint! — Function

add_joint!(mechanism, jointdata; [parent, child, jointID, check_frames_exist::Bool])

Add a joint to a mechanism. Returns the joint id.

Arguments

jointID: the name of the joint, a String. If not provided, a unique jointID is created.

parent: the name of the parent frame, a String. If not provided, then the default is the most recently added frame.

child the name of the child frame, a String. If not provided, adds a new frame to the robot.

check_frames_exist: if true, then check that the parent and child frame already exist in the mechanism. If false, then this check is skipped. In this case the frames are not added to the mechanism.

See Also

To add a frame to a mechanism see add_frame!. For more information on types of joint see AbstractJointData and MechanismJoint.

VMRobotControl.add_coordinate! — Function

add_coordinate!(m::Mechanism, c::CoordinateData; id::String)
add_coordinate!(m::VirtualMechanismSystem, c::CoordinateData; id::String)

Add a coordinate to a mechanism or virtual mechanism system. The coordID is the name of the coordinate.

See also get_compiled_coordID, configuration, velocity, acceleration, jacobian

VMRobotControl.add_component! — Function

add_component!(m::Mechanism, c::ComponentData; id::String)
add_component!(m::VirtualMechanismSystem, c::ComponentData; id::String)

Add a coordinate to a mechanism or virtual mechanism system. The coordID is the name of the coordinate.

See also get_compiled_componentID

VMRobotControl.add_inertia! — Function

add_inertia!(mechanism, frameID, inertia; id, coordID=nothing)

Add an inertia to a mechanism. A FrameAngularVelocity coordinate is added to the mechanism, on the frame with ID frameID, and is given the ID coordID if provided.

The inertia is defined by a 3x3 matrix inertia.

VMRobotControl.add_gravity_compensation! — Function

add_gravity_compensation!(system, gravity)

Add gravity compensators for all point masses.

If system is a VirtualMechanismSystem, then only point masses in the robot are compensated, and the GravityCompensator components are added to the system, so that they are treated as part of the controller, not part of the robot.

VMRobotControl.add_gravity_compensator! — Function

add_gravity_compensator!(mechanism, point_mass, gravity; id=nothing)

Add a gravity compensator to componsate for point_mass, when subject to gravity.

Using

Compiling and Cache-bundles

VMRobotControl.compile — Function

compile(m::Mechanism)
compile(m::VirtualMechanismSystem)

Compile a Mechanism or VirtualMechanismSystem into an immutable form used for computations.

VMRobotControl.new_kinematics_cache — Function

new_kinematics_cache(m[, T])

Creates a cache with element type ::T (defaulting to the eltype of the mechanism) for performing kinematics computations for a compiled mechanism or compiled virtual mechanism system.

This cache supports computation of frame transforms and coordinate configuration's, but do NOT support computing jacobians or velocities.

This type of cache is commonly used for plotting, where only the frame transforms are needed.

See also kinematics!, new_jacobians_cache, new_rbstates_cache, new_dynamics_cache,

VMRobotControl.new_jacobians_cache — Function

new_jacobians_cache(m[, T])

Creates a cache with element type ::T (defaulting to the eltype of the mechanism) for performing kinematics and jacobian computations for a compiled mechanism or compiled virtual mechanism system.

This cache supports computation of frame transforms and jacobians, configuration and jacobian for coordinates, but do NOT support velocities.

See also jacobians!, new_rbstates_cache, new_dynamics_cache,

VMRobotControl.new_rbstates_cache — Function

new_rbstates_cache(m[, T])

Creates a cache with element type ::T (defaulting to the eltype of the mechanism) for performing kinematics and velocity computations for a compiled mechanism or compiled virtual mechanism system.

This cache supports computation of frame transforms, velocities and accelerations. Depending on which computation is performed, velocity_kinematics! or acceleration_kinematics!, the acceleration will either be a true acceleration dependant upon joint accelerations q̈, or the velocity product acceleration, which is the acceleration assuming q̈ = 0. It supports configuration, velocity, and acceleration, for coordinates. It does NOT support jacobians or generalized forces/torques.

See also velocity_kinematics! new_dynamics_cache,

VMRobotControl.new_dynamics_cache — Function

new_dynamics_cache(m[, T])

Creates a cache with element type ::T (defaulting to the eltype of the mechanism) for performing kinematics, jacobian, and velocity computations for a compiled mechanism or compiled virtual mechanism system, as well as generalized forces/torques from components, the inertance matrix, of the system, and the acceleration of the system, q̈.

This cache supports computation of frame transforms, velocities, velocity-product-accelerations, and jacobians. It also supports configuration, velocity, acceleration, and jacobian for coordinates.

See also dynamics!

VMRobotControl.new_control_cache — Function

new_control_cache(m[, T])

Creates a cache with element type ::T (defaulting to the eltype of the mechanism) for controlling a robot with a virtual mechanism controller.

Note

Components attached to the robot will NOT be used for control. The components on the robot represent the model of the robot, and are not considered part of the controller. If you wish to control the robot, you must add components to the VirtualMechanismSystem. If your component uses a coordinate from the robot, you can add it to the VirtualMechanismSystem and refer to the coordinate using the id ".robot.<coord_id>"

See also control_step!

Accessing data from cache-bundles

VMRobotControl.get_rbstate — Function

get_rbstate(cache, frame::CompiledFrameID)

Returns the rigid body state of frame from the cache.

VMRobotControl.get_transform — Function

get_transform(cache, frame::CompiledFrameID)

Returns the transform of the frame from the cache.

VMRobotControl.get_linear_vel — Function

get_linear_vel(cache, frame::CompiledFrameID)

Returns the linear velocity of the origin of frame from the cache.

VMRobotControl.get_angular_vel — Function

get_angular_vel(cache, frame::CompiledFrameID)

Returns the angular velocity of the frame from the cache.

VMRobotControl.get_linear_vpa — Function

get_linear_vpa(cache, frame::CompiledFrameID)

Returns the linear velocity product acceleration of the origin of frame from the cache.

VMRobotControl.get_angular_vpa — Function

get_angular_vpa(cache, frame::CompiledFrameID)

Returns the angular velocity product acceleration of frame from the cache.

VMRobotControl.get_linear_acceleration — Function

get_linear_acceleration(cache, frame::CompiledFrameID)

Returns the total linear acceleration of the origin of frame from the cache.

VMRobotControl.get_angular_acceleration — Function

get_angular_acceleration(cache, frame::CompiledFrameID)

Returns the total angular acceleration of frame from the cache.

VMRobotControl.get_linear_jacobian — Function

get_linear_jacobian(cache, frame::CompiledFrameID)

Returns the linear Jacobian of frame from the cache.

VMRobotControl.get_angular_jacobian — Function

get_angular_jacobian(cache, frame::CompiledFrameID)

Returns the angular Jacobian of frame from the cache.

VMRobotControl.get_t — Function

get_t(cache)

Returns the time from the cache.

VMRobotControl.get_q — Function

get_q(cache)

Returns the joint configurations from the cache.

VMRobotControl.get_q̇ — Function

get_q̇(cache)

Returns the joint velocities from the cache.

VMRobotControl.get_q̈ — Function

get_q̈(cache)

Returns the joint accelerations from the cache.

VMRobotControl.get_u — Function

get_u(cache)

Returns the additional joint torques from the cache.

This can be used during f_control to apply additional torques to the system, by modifying the returned torques inplace. By default, these torques are zero until set otherwise.

VMRobotControl.get_inertance_matrix — Function

get_inertance_matrix(cache)

Returns the inertance matrix in the cache.

For a virtual mechanism system, this is a tuple of inertance matrices: one for the robot and one for the virtual mechanism.

VMRobotControl.get_generalized_force — Function

get_generalized_force(cache)

Returns the generalized forces in the cache.

For a virtual mechanism system, this is a tuple of vectors: one for the robot and one for the virtual mechanism.

Computations

VMRobotControl.kinematics! — Function

kinematics!(bundle, t, q)

Compute the kinematics/transforms for the system.

Arguments

t time

q joint configuration vector

See also get_transform

VMRobotControl.jacobians! — Function

jacobians!(bundle, t, q)

Compute the Jacobians for the mechanism at time t and configuration q and store them in cache.jacobians.

VMRobotControl.velocity_kinematics! — Function

velocity_kinematics!(cache, t, q, q̇)

Compute the velocities of the system at time t and configuration q, with velocities q̇, and store them in cache.rbstates. Also computes the velocity product accelerations (i.e. accelerations assuming q̈ = 0).

Also computes the kinematics as a prerequisite.

VMRobotControl.acceleration_kinematics! — Function

acceleration_kinematics!(bundle, t, q, q̇, q̈)

Compute the accelerations for the mechanism with known joint-accelerations q̈.

Also computes the kinematics/velocities as prerequisites.

Arguments

t time

q joint configuration vector

q̇ joint velocity vector

q̈ joint acceleration vector

VMRobotControl.precompute! — Function

precompute!(cache, t, q, q̇, gravity, u)

Precompute the necessary quantities for the dynamics of the mechanism at time t and configuration q, with velocities q̇, gravitational acceleration gravity, and generalized forces u.

This includes the velocity kinematics and Jacobians, and stores the results in cache. The accelerations q̈ are not computed.

VMRobotControl.generalized_force! — Function

generalized_force!(τ, bundle::CacheBundle, c::ComponentData)

Using the current state of the the cache, calculates the generalized force due to component c, and adds it to τ.

Typically τ is a vector from the cache, but it is kept as an argument to allow for more flexibility in the use of this function.

VMRobotControl.inertance_matrix! — Function

inertance_matrix!(M, bundle::CacheBundle, c::ComponentData)

Using the current state of the cache, calculates the inertance matrix due to component c, and adds it to M.

Typically M is a matrix from the cache, but it is kept as an argument to allow for more flexibility in the use of this function.

VMRobotControl.dynamics! — Function

dynamics!(cache, t, q, q̇, gravity, [u])

Compute the accelerations for the mechanism at time t and configuration q, with velocities q̇, gravitational acceleration gravity, and generalized forces u, and store the results in the cache.

If u is not provided, it is assumed to be zero.

VMRobotControl.control_step! — Function

control_step!(bundle, t, qʳ, q̇ʳ)

Perform a step of the control loop for the system.

Inputs are the current time t, the robot joint configuration qʳ, and the robot joint velocity q̇ʳ.

The elapsed time dt is computed from the previous time and t. The virtual mechanism state is updated using a forward Euler step with the computed acceleration q̈ᵛ and the elapsed time dt.

Note

Components attached to the robot will NOT be used for control. The components on the robot represent the model of the robot, and are not considered part of the controller. If you wish to control the robot, you must add components to the VirtualMechanismSystem. If your component uses a coordinate from the robot, you can add it to the VirtualMechanismSystem and refer to the coordinate using the id ".robot.<coord_id>"

Returns the torques to be applied to the robot.

See also control_step!

VMRobotControl.stored_energy — Function

stored_energy(bundle::MechanismCacheBundle, c::ComponentData)

Computes the stored energy of component c, kinetic and potential.

VMRobotControl.opspace_force — Function

opspace_force(bundle::MechanismCacheBundle, c::ComponentData)

Calulates and returns the operational space force due to component c.

Accessing compiled frames/coords/joints/components

VMRobotControl.get_compiled_frameID — Function

get_compiled_frameID(m::Union{CompiledMechanism, CompiledVirtualMechanismSystem}, c::String)

Returns compiled frame ID of frame with name c. This can be used to get the frames's transform, velocity, jacobian etc. from a kinematics/dynamics cache.

VMRobotControl.get_compiled_coordID — Function

get_compiled_coordID(m::Union{CompiledMechanism, CompiledVirtualMechanismSystem}, c::String)

Returns the compiled coordinate ID of a coordinate with name c. This can be used to get the configuration/velocity/jacobian of a coordinate from a kinematics/dynamics cache.

VMRobotControl.get_compiled_jointID — Function

get_compiled_jointID(m::Union{CompiledMechanism, CompiledVirtualMechanismSystem}, c::String)

Returns the compiled joint ID of a joint with name c. This can be used to change the joint data for a compiled system, e.g. when numerically optimizing the robot structure.

VMRobotControl.get_compiled_componentID — Function

get_compiled_componentID(m::Union{CompiledMechanism, CompiledVirtualMechanismSystem}, c::String)

Returns the compiled component ID of a component with name c. Can be used to mutate the component data of a compiled mechanism, or recompute information such as the force generated by the component.

Setting up simulations

VMRobotControl.get_ode_dynamics — Function

get_ode_dynamics(cache, gravity; [f_setup, f_control])

Returns a function for the autonomous ODE dynamics of the mechanism in the form ̇x = f(x, p, t) where the parameters p are ignored.

This form is used for DifferentialEquations.jl

f_setup, f_control, and f_shutdown

The optional keyword arguments f_setup and f_control are functions that can be used to implement custom behaviour.

f_setup(cache) is called once before starting.

f_control(cache, t, setup_ret, extra) is called at each time step to allow for custom control logic. setup_ret is the return value of f_setup, and can be used to initialize variables before the controller starts. You can directly modify the robot/virtual mechanism state in this function, or modify the vectors returned by get_u(cache) to apply additional actuation to the robot or virtual mechanism.

VMRobotControl.assemble_state — Function

assemble_state(m::Union{CompiledMechanism, CompiledVirtualMechanismSystem}, q, q̇)

Assembles a state vector from a configuration q and velocity q̇. This is useful for setting up ODEs for simulation.

VMRobotControl.get_ode_problem — Function

get_ode_problem(cache, gravity, q, q̇, tspan, [p, f_setup, f_control]; kwargs...)

Construct an ODE problem. See also VMRobotControl.get_ode_dynamics. initial conditions are given by q and q̇.

DifferentialEquations.jl must be loaded for this method to be defined: using DifferentialEquations Implementations are in in ./ext/VMRobotControlDifferentialEquationsExt.jl

p and kwargs are optional arguments that can be used to customize the ODEProblem. f_setup and f_control are optional arguments that can be used to customize the dynamics function.

VMRobotControl.zero_q — Function

zero_q(m::Union{CompiledMechanism, CompiledVirtualMechanismSystem})

Returns a zero configuration for the system. For a mechanism, this is a vector of zeros, and for a virtual mechanism system, this is a tuple of vectors of zeros (one for the robot, one for the virtual mechanism).

VMRobotControl.zero_q̇ — Function

zero_q̇(m::Union{CompiledMechanism, CompiledVirtualMechanismSystem})

Returns a zero velocity for the system. For a mechanism, this is a vector of zeros, and for a virtual mechanism system, this is a tuple of vectors of zeros (one for the robot, one for the virtual mechanism).

VMRobotControl.zero_q̈ — Function

zero_q̈(m::Union{CompiledMechanism, CompiledVirtualMechanismSystem})

Returns a zero acceleration for the system. For a mechanism, this is a vector of zeros, and for a virtual mechanism system, this is a tuple of vectors of zeros (one for the robot, one for the virtual mechanism).

VMRobotControl.zero_u — Function

zero_u(m::Union{CompiledMechanism, CompiledVirtualMechanismSystem})

Returns a zero control input for the system. For a mechanism, this is a vector of zeros, and for a virtual mechanism system, this is a tuple of vectors of zeros (one for the robot, one for the virtual mechanism).

Plotting

Joints

AbstractJointData types

VMRobotControl.Rigid — Type

Rigid(::Transform)

A rigid joint, defined simply by a transformation from the child frame to the parent frame.

VMRobotControl.Revolute — Function

Revolute(axis::SVector{3})
Revolute(axis::SVector{3}, tf::Transform)

Represents a revolute/rotating joint between two frames. The transform from the child frame to the parent frame is tf * Transform(AxisAngle(axis, q)).

axis must be normalized: norm(axis)≈1.0.

See also Prismatic, Rail

VMRobotControl.Prismatic — Function

Prismatic(axis::SVector{3})
Prismatic(axis::SVector{3}, tf::Transform)

Represents a prismatic/sliding joint between two frames. The transform from the child frame to the parent frame is tf * Transform(q*axis). The axis does not have to be normalized.

See also Revolute, Rail,

VMRobotControl.Rail — Function

Rail(spline, tf::Transform)
Rail(spline)

Represents a prismatic/sliding joint where the joint slides along a bezier spline. The child frame is not rotated relative to the parent frame, (unless tf causes a rotation).

The origin of the parent frame for joint variable q is tf * spline_position(spline, q).

See also Prismatic, Revolute

VMRobotControl.ReferenceJoint — Type

ReferenceJoint(::Transform)

A ReferenceJoint does not exhibit dynamics, but can it's transform can be mutated, so that it may act as a motion source. The transform is stored as A RefValue, which you can set/getcan like: joint.transform[].

Be careful when changing the transform–-if you are doing many simulations you may need to deepcopy to avoid aliasing.

Coordinates

VMRobotControl.CoordinateData — Type

CoordinateData

An abstract type representing an operation space coordinate.

See also: configuration, velocity, jacobian and acceleration

Coordinate Interface

VMRobotControl.configuration — Function

configuration(cache, coordinate)

Returns the configuration of the operation space coordinate coordinate z.

VMRobotControl.velocity — Function

velocity(cache, coordinate)

Returns the velocity of the operation space coordinate coordinate ż, the derivative of the configuration.

VMRobotControl.jacobian — Function

jacobian(cache, coordinate)

Return the jacobian of the operation space coordinate coordinate J, such that ż = Jq̇, where ż is the opspaceVelocity and q̇ are the joint velocities.

Missing docstring.

Missing docstring for vpa. Check Documenter's build log for details.

Coordinate types

VMRobotControl.CoordDifference — Type

CoordDifference(parent, child)
CoordDifference(;parent, child)

Represents the difference between two coordinates. For example if both coordinates are cartesian postions, then this represents the vector from one to the other.

VMRobotControl.CoordStack — Type

CoordStack(c1, c2)
CoordStack(;c1, c2)

Represents the vertical concatenation of two coordinates.

VMRobotControl.CoordSlice — Type

CoordSlice(coordid, idxs::SVector{N, Int})

A coordinate made by indexing into another coordinate.

VMRobotControl.CoordSum — Type

CoordSum(c1, c2)
CoordSum(;c1, 1child)

Represents the sum of two coordinates.

VMRobotControl.ConstCoord — Type

ConstCoord(::SVector)

A constant coordinate. Can be used with coordinate differences to shift the equilibrium of a spring.

VMRobotControl.JointSubspace — Type

JointSubspace(jointid)

Represents the configuration of the joint identified by jointid.

When compiled, is replaced with the compiled mechanism joint representation. This is because the joint type is needed to know the size of the joint.

VMRobotControl.FrameOrigin — Type

FrameOrigin(frameID)

The position of the origin of frame frameID in the world.

VMRobotControl.FramePoint — Type

FramePoint(frameID, point::SVector{3})

A point fixed in the frame designated by frameID. point is the position of the point in the frame. The configuration of this coordinate is the vector to the point, represented in the root frame of the mechanism.

VMRobotControl.FrameAngularVelocity — Type

FrameAngularVelocity(frameID)

The angular velocity for frame frameID in world frame. Used to implement rotational inertia components..

VMRobotControl.QuaternionAttitude — Type

QuaternionAttitude(frameID)

The rotation for frame frameID in the world, as the vector part of the the quaternion. Can be combined with a linear spring for continuous attitude control.

See "Quaternion-based impedance with nondiagonal stiffness for robot manipulators" by Fabrizio Caccavale et al, 1998 IEEE ACC.

VMRobotControl.ReferenceCoord — Type

ReferenceCoord(::Base.RefValue{SVector{N, T}})

A reference coordinate. By keeping a reference to a coordinate, it can be modified in place.

VMRobotControl.RotatedCoord — Type

RotatedCoord(worldframecoord, frameID)

Rotates worldframecoordinate z⁽ʷ⁾ to the frame of link designated by frameID, such that z = R⁻¹ z⁽ʷ⁾.

VMRobotControl.UnrotatedCoord — Type

UnrotatedCoord(linkframecoord, frameID)

Rotates link frame coordinate z⁽ˡ⁾ from the frame of link designated by frameID, to the world frame, so that z = Rʷˡ z⁽ˡ⁾.

Components

VMRobotControl.Storage — Type

Storage{T} <: ComponentData{T}

Supertype for components that act as potential energy storages, with float type T.

These must support implement stored_energy and generalized_force!, typically by implementing _stored_energy and _opspace_force functions, respectively.

VMRobotControl.LinearSpring — Type

LinearSpring(stiffness, coord::OperationSpace)

A linear spring acting on the provided coordinate. The spring stiffness stiffness can be any type that supports multiplication with the coordinate's configuration, but should be positive definite to ensure it's energy is bounded below, and should not cause allocations when multiplied (i.e. use a Float64 or an SMatrix).

VMRobotControl.TanhSpring — Type

TanhSpring(coord::OperationSpace; stiffness, max_force)

A spring which applies a force of magnitude tanh(|extension|) acting on the provided coordinate. The spring can be configured in terms of 2 of 3 variables: either a width a max_force, or a stiffness, which are each scalar.

The force is determined as: F = kβ tanh(|z|/β) z/|z| where z is the coordinate's configuration, β is the width of the spring, and kis the stiffness. The max force σ iskβ`.

This is useful as the tanh function saturates to a constant force for large extensions, which can prevent the robot from applying too much force.

VMRobotControl.GaussianSpring — Type

GaussianSpring(σ, stiffness, coordid)
GaussianSpring(coordid; stiffness, max_force, width)

A spring with a Gaussian energy profile, which is useful for acting as a localized attractive/repulsive force (repulsive when stiffness is negative).

The energy function is a gaussian with std deviation σ, scaled so that the gradient of it's derivative (the force) is k at |z|=0: E = -kσ² exp( -|z|²/(2σ²) ) where z is the coordinate's configuration.

When calling with keyword arguments, specify any two, as this fully defines the Gaussian energy function.

VMRobotControl.GravityCompensator — Type

GravityCompensator(mass::T, coord::OperationSpace{3}, gravity::SVector{3, T})
GravityCompensator(;mass, coord, gravity=DEFAULT_GRAVITY)

An energy storage acting in the opposite direction to the vector gravity. Can be used to compensate for the potential energy of a pointmass on the same coordinate, without affecting inertia.

Note that the stored energy is potentially unbounded for robots with unbounded workspace, but is reported as the potential energy of the mass at the current configuration.

See also add_gravity_compensator!, add_gravity_compensation!.

VMRobotControl.ForceSource — Type

ForceSource(force_max, power_max, coord)

A power source which applies force up to force_max in the direction of force_max maximum with power power_max. The force is applied at coord.

If power max is a scalar, then it represents a maximum power flow into the system. If it is a tuple of two scalars, then the first element bounds power flow into the system, and the second element bounds the power flow out of the system.

VMRobotControl.PowerSource — Type

PowerSource(force_max, power_max, coord)

A power source which applies a force up to force_max in the direction that will inject power into the system up to power_max. Cannot extract power from the system. The force is applied at coord.

VMRobotControl.LinearDamper — Type

LinearDamper(damping, coord::OperationSpace)

A linear damper acting on the provided coordinate. The damping coefficient damping can be any type that supports multiplication with the coordinate's velocity, but should be positive definite to ensure it is dissipative, and should not cause allocations when multiplied (i.e. use a Float64 or an SMatrix).

VMRobotControl.TanhDamper — Type

TanhDamper(max_force, width, coord::OperationSpace)

A damper that applies a constant damping force above a certain velocity, using a tanh function. The width parameter determines the shape of the tanh, at 2width, the damping force is ~95% of the maximum, and at 3width, the damping force is ~99.5% of the maximum.

VMRobotControl.RectifiedDamper — Type

RectifiedDamper(damping, coord::OperationSpace, bounds, flipped, diodic)

A linear damper acting on the provided coordinate only in a certain position range. If position z is greater than ub, full damping is applied, if it is between the lower bound and upper bound, partial damping is applied, and if it is less than the lower bound, no damping is applied.

If flipped is true, this is inverted and the damper is applied when z is less than the lower bound, and not applied when z is greater than the upper.

If diodic is true, the damper is applied only when the velocity is positive (or negative if flipped is true).

VMRobotControl.DiodicDamper — Type

DiodicDamper(damping, coord::OperationSpace, flipped)

A that acts only when the velocity is positive (or negative if flipped is true).

VMRobotControl.LinearInerter — Type

LinearInerter(inertance, coord)

A linear inertance with inertance matrix inertance across coordinate coord.

The inertance matrix should be positive definite, and should multiply with the velocity vector to produce a force vector, or the jacobian to produce an inertance matrix. It should not cause allocation when multiplied (i.e use SMatrix or Float64).

A linear inerter will have the same inerting effect as a point mass if the inertance equals the mass, and the coordinate is the center of mass, but is not subject to gravity.

VMRobotControl.Inertia — Type

Inertia(inertia::SMatrix{3, 3}, coord)

A rotational inertia. Coord must be a FrameAngularVelocity coordinate.

See also add_inertia!.

VMRobotControl.PointMass — Type

PointMass(mass, coord_id)

A point mass with mass mass at coordinate coord. The coord should be a 3 dimensional in the global frame, such as a FrameOrigin or a FramePoint.

The point mass is subject to gravity. This can be compensated using a GravityCompensator component, or avoided by using a LinearInerter instead of a point mass.