Additional info for Developers

VMRobotControl.AbstractJointData

VMRobotControl.CompiledCoordID

VMRobotControl.CompiledMechanism

VMRobotControl.CompiledVirtualMechanismSystem

VMRobotControl.ComponentData

VMRobotControl.ConstCoord

VMRobotControl.CoordDifference

VMRobotControl.CoordSlice

VMRobotControl.CoordStack

VMRobotControl.CoordSum

VMRobotControl.CoordinateData

VMRobotControl.DiodicDamper

VMRobotControl.Dissipation

VMRobotControl.EltypeCache

VMRobotControl.ForceSource

VMRobotControl.FrameAngularVelocity

VMRobotControl.FrameOrigin

VMRobotControl.FramePoint

VMRobotControl.GaussianSpring

VMRobotControl.GenericComponent

VMRobotControl.GravityCompensator

VMRobotControl.Inertance

VMRobotControl.Inertia

VMRobotControl.JointSubspace

VMRobotControl.LinearDamper

VMRobotControl.LinearInerter

VMRobotControl.LinearSpring

VMRobotControl.Mechanism

VMRobotControl.MechanismJoint

VMRobotControl.PointMass

VMRobotControl.PowerSource

VMRobotControl.QuaternionAttitude

VMRobotControl.RBTree

VMRobotControl.RSON.RSONParserConfig

VMRobotControl.RectifiedDamper

VMRobotControl.ReferenceCoord

VMRobotControl.ReferenceJoint

VMRobotControl.Rigid

VMRobotControl.RotatedCoord

VMRobotControl.Storage

VMRobotControl.TanhDamper

VMRobotControl.TanhSpring

VMRobotControl.TimeFuncJoint

VMRobotControl.Transforms.EulerVector

VMRobotControl.Transforms.SpatialAcceleration

VMRobotControl.Transforms.Transform

VMRobotControl.Transforms.Twist

VMRobotControl.TypeStableCollections.TypeStableCollection

VMRobotControl.TypeStableCollections.TypeStableCollection

VMRobotControl.TypeStableCollections.TypeStableIdx

VMRobotControl.URDF.URDFParserConfig

VMRobotControl.UnrotatedCoord

VMRobotControl.VirtualMechanismSystem

VMRobotControl.Prismatic

VMRobotControl.RSON.parseRSON

VMRobotControl.RSON.parseRSONString

VMRobotControl.RSON.parseRSONVersion

VMRobotControl.RSON.serializeRSON

VMRobotControl.Rail

VMRobotControl.Revolute

VMRobotControl.Spherical

VMRobotControl.Splines.curve_derivative

VMRobotControl.Splines.curve_position

VMRobotControl.Splines.curve_second_derivative

VMRobotControl.Transforms.AxisAngleDerivative

VMRobotControl.Transforms.angular_velocity_prematrix

VMRobotControl.Transforms.angular_velocity_prematrix_derivative

VMRobotControl.Transforms.quaternion_derivative_propagation

VMRobotControl.Transforms.quaternion_derivative_propagation_derivative

VMRobotControl.Transforms.quatmul_geodual_bivector_matrix

VMRobotControl.Transforms.skew

VMRobotControl.TypeStableCollections.extend

VMRobotControl.TypeStableCollections.push_return_index!

VMRobotControl.URDF.parseURDF

VMRobotControl.URDF.parseURDFString

VMRobotControl._add_jacobian_transpose_times_force!

VMRobotControl._construct_coordinate_graph

VMRobotControl.acceleration_kinematics!

VMRobotControl.add_component!

VMRobotControl.add_coordinate!

VMRobotControl.add_frame!

VMRobotControl.add_gravity_compensation!

VMRobotControl.add_gravity_compensator!

VMRobotControl.add_inertia!

VMRobotControl.add_joint!

VMRobotControl.assemble_state

VMRobotControl.autodiff_jacobian_column

VMRobotControl.child_frame_twist

VMRobotControl.child_rbstate

VMRobotControl.compile

VMRobotControl.components

VMRobotControl.config_size

VMRobotControl.configuration

VMRobotControl.control_step!

VMRobotControl.coordinates

VMRobotControl.dissipations

VMRobotControl.dynamics!

VMRobotControl.frame_ancestor_matrix

VMRobotControl.frames

VMRobotControl.generalized_force!

VMRobotControl.generic_components

VMRobotControl.get_angular_acceleration

VMRobotControl.get_angular_jacobian

VMRobotControl.get_angular_vel

VMRobotControl.get_angular_vpa

VMRobotControl.get_compiled_componentID

VMRobotControl.get_compiled_coordID

VMRobotControl.get_compiled_frameID

VMRobotControl.get_compiled_jointID

VMRobotControl.get_generalized_force

VMRobotControl.get_inertance_matrix

VMRobotControl.get_linear_acceleration

VMRobotControl.get_linear_jacobian

VMRobotControl.get_linear_vel

VMRobotControl.get_linear_vpa

VMRobotControl.get_ode_dynamics

VMRobotControl.get_ode_problem

VMRobotControl.get_q

VMRobotControl.get_q̇

VMRobotControl.get_q̈

VMRobotControl.get_rbstate

VMRobotControl.get_t

VMRobotControl.get_transform

VMRobotControl.get_u

VMRobotControl.inertance_matrix!

VMRobotControl.inertances

VMRobotControl.jacobian

VMRobotControl.jacobian_column

VMRobotControl.jacobians!

VMRobotControl.joint_relative_acceleration

VMRobotControl.joint_relative_transform

VMRobotControl.joint_relative_twist

VMRobotControl.joint_relative_vpa

VMRobotControl.joints

VMRobotControl.kinematics!

VMRobotControl.name

VMRobotControl.ndof

VMRobotControl.new_control_cache

VMRobotControl.new_dynamics_cache

VMRobotControl.new_jacobians_cache

VMRobotControl.new_kinematics_cache

VMRobotControl.new_rbstates_cache

VMRobotControl.num_frames

VMRobotControl.opspace_force

VMRobotControl.precompute!

VMRobotControl.reassign_coords

VMRobotControl.reassign_frames

VMRobotControl.reassign_joints

VMRobotControl.remake

VMRobotControl.rigid_joint_twist

VMRobotControl.root_frame

VMRobotControl.state_idxs

VMRobotControl.storages

VMRobotControl.stored_energy

VMRobotControl.velocity

VMRobotControl.velocity_kinematics!

VMRobotControl.velocity_size

VMRobotControl.visuals

VMRobotControl.zero_q

VMRobotControl.zero_q̇

VMRobotControl.zero_q̈

VMRobotControl.zero_u

Testing

To test, navigate to the folder "VMRobotControl.jl", hit "]" to enter the Pkg context, and run

test

Kinematics computations

VMRobotControl.RBTree — Type

RBTree(m:Mechanism)

Fundamental data structure behind efficient kinematics computations. Used to store caches for computations, and information about compiled frames/joints/coordinates

Strategy:

There are a few things to do when going from a Mechanism to an CompiledMechanism. The first is to change all frames from being referenced by strings to being referenced by indexes (of type Int). This is done using the reassign_frames function for coordinates/components and in the constructor for joints.

The second is to change references to joints from strings of the joints name to a suitable alternative, like a compiled joint ID, as the size of the joint-configuration vector q must be able to be determined by the type of the ID in order to know it's size at compile time. This is done using the reassign_joints

The next is to create a type stable way to compute the extended kinematics (transforms, twists and velocity product accelerations of each frame), and the jacobians of each frame. This is accomplished by building a type-agnostic cache for each, and creating a Walk of the rigid body tree. The walk is simply a concrete order in which to compute and apply joint transformation. For example, the tree with frames (1, 2, 3, 4), and structure: 1 -> 2 -> 3 -> 4 Could have a walk [(1, 2), (2, 3), (1, 4)], or equivalently [(1, 2), (1, 4), (2, 3)], or [(1, 4), (1, 2), (2, 3)]. The important fact is that a transform earlier in the tree is always computed before a transform later in the tree.

Each different joint type requires different methods, we create a joint_collection which has a field for each different joint type. This allows us to use map or foreach to efficiently dispatch to joints to compute relative transforms/twists/vpas, and then we use the walk to accumulate the effect of each. This works because despite each joint being a different type, the output of map or foreach is a the same.

For coordinates, a different approach is needed to store the dependency order. First we determine a suitable TypeStableCollection type for the collection of all coordinates, then we create a directed graph of dependencies between coordinates. We then topologically sort the graph to determine the order in which to compute the coordinates. This is done by first computing the depth of each coordinate, and then creating a vector of vectors where each inner vector is a set of coordinates that can be computed simultaneously, as they do not depend on each other.

VMRobotControl.joint_relative_transform — Function

joint_relative_transform(joint, t, q)

Returns the transform from the child frame to the parent frame of a joint configuration and time.

The transform is represented as a Transform object

q is the joint configuration for the individual joint (i.e. not for the whole mechanism) as an SVector of the appropriate size.

This function is used to ensure correct type signature for all joints and arguments, while implementation for each AbstractJointData is done in a method of joint_relative_transform.

VMRobotControl.joint_relative_twist — Function

joint_relative_twist(joint, t, q, q̇)

Returns the twist of the child frame of a joint represented in the parent frame.

The twist is represented as a Twist object.

q is the joint configuration for the individual joint (i.e. not for the whole mechanism) as an SVector of the appropriate size.

q̇ is the joint velocity for the individual joint (i.e. not for the whole mechanism) as an SVector of the appropriate size.

VMRobotControl.child_frame_twist — Function

child_frame_twist(tf_parent, tf_child, twist_parent, joint_relative_twist)

Compute the twist of the child frame, from the twist of the parent frame, the transforms (from each frame to the world frame) and the joint relative twist: i.e. the twist of the child frame represented in the parent frame.

VMRobotControl.joint_relative_vpa — Function

joint_relative_vpa(joint, t, q, q̇)

Returns the velocity-product-acceleration of the child frame of a joint represented in the parent frame.

The function joint_vpa applies the correction due to velocity of the parent frame.

The acceleration is represented as a SpatialAcceleration object

Most joint types have no velocity-product-acceleration, and return zero.

What is velocity-product-acceleration?

Velocity product acceleration is the acceleration of the child frame relative to the parent frame due to the velocity of the joint variable, q̇, assuming q̈=0, and that the parent frame is stationary. As an example, consider the linear velocity as ȯ = J(q)q̇, then the derivative with time is ö = J(q)q̈ + J̇(q)q̇. Here J(q) is the jacobian of the joint. Thus J̇(q) ̇q is the velocity product acceleration.

VMRobotControl.joint_relative_acceleration — Function

joint_relative_acceleration(joint, q, q̈, t)

Returns the acceleration of the child frame relative to the parent frame due to the acceleration of the joint variable, q̈.

Does NOT include velocity-product-accelerations terms like centrifugal/coriolis acceleration.

VMRobotControl.autodiff_jacobian_column — Function

autodiff_jacobian_column(joint1dof, tf_parent, tf_child, tf_final, qi)

Used to efficiently compute jacobians using automatic differentiation.

Uses precomputed transforms to compute the derivative of a final transform tf_final w.r.t. a single joint config, qi, by using knowledge that the other transforms are not a function of qi

VMRobotControl.child_rbstate — Function

child_rbstate(state_parent, state_relative)

Computes the rigid body state of a child from from the parent state, and the relative state. state_parent is the state of the parent rigid body (transform, twist, vpa). state_relative contains the relative transform/twist/vpa between the child frame and parent frame.

TypeStableCollection

Let me illustrate the challenge: you have a robot with a mix of types of joint: revolute, prismatic and potentially other types too. You want to write a function that can take any such robot and compute its kinematics. If you simply make a list of joints, then it will not be concretely typed (e.g. `Vector{Any}. It is possible that small union optimisations would work for this example, but that doesn't scale to the number of different coordinate types or component types.)

Instead, we could use a tuple of the joints, but this would mean a different type for the robot for every arrangement of joints. Instead we use a tuple of vectors: one vector for each different type. This means that any robot made of only revolute and prismatic joints will be of the same type! We store the order of computation separately, by storing indices into the tuple of vectors.

The tuple-of-vectors structure is called a TypeStableCollection and has drawn inspiration from TypeSortedCollection, but has some difference in construction and indexing.

VMRobotControl.TypeStableCollections.TypeStableCollection — Type

TypeStableCollection{TUPTYPE}

A collection of vectors, where each vector has a different element type. The element types are stored in a tuple of type TUPTYPE.

Allows for type stable operations on the collection, such as map, mapreduce, foreach.

Can be built using the extend function to add new/existing types and return an index; [push!] to add types already present; or push_return_index! to add types already present and return the index of the newly added element.

VMRobotControl.TypeStableCollections.TypeStableCollection — Method

TypeStableCollection()

Create an empty TypeStableCollection, useful to then use extend.

VMRobotControl.TypeStableCollections.TypeStableIdx — Type

TypeStableIdx{T}

An index into a TypeStableCollection.

The index is parameterized by the type of the element it points to, T. Therefore it can still be used to access an extended collection, even if the type of the collection has changed, or if elements have been appended to the collection.

VMRobotControl.TypeStableCollections.extend — Method

extend(c::TypeStableCollection, val)

Extend the type stable collection by adding val. If the type of val is already in the collection, val is appended to the corresponding vector. If the type is not already in the collection, a new vector is created, thus changing the type of the collection.

As a result, the type of the collection may change after calling extend. Therefore, to keep the calling code type stable, it is recommended to always name the return value of extend with a new variable, and use that variable in subsequent code.

Returns: the new collection and a TypeStableIdx to the newly added value.

Examples

julia> c1 = TypeStableCollection(())
TypeStableCollection{Tuple{}}(())
julia> c2, idx = extend(c1, 2)

VMRobotControl.TypeStableCollections.push_return_index! — Method

push_return_index!(c::TypeStableCollection, val)

Push val to the collection c. Requires that type of val is already in the collection.

Returns: the modified collection and a TypeStableIdx to the newly added value.

Transforms

VMRobotControl.Transforms.EulerVector — Type

EulerVector(v::SVector{3})

Represents a rotation on the plane with normal vector v, by angle norm(v).

VMRobotControl.Transforms.SpatialAcceleration — Type

SpatialAcceleration{T}

A spatial acceleration represents the acceleration of a rigid body/frame in 3D space.

It is composed of a linear acceleration and an angular acceleration.

VMRobotControl.Transforms.Transform — Type

Transform{T}(origin::SVector{3, T}, rotor::Rotor{T})

A rigid body transform defined by an origin and a rotor (quaternion).

Can be constructed from either an origin and a rotor, or just an origin, or just a rotor.

VMRobotControl.Transforms.Twist — Type

Twist{T}(linear::SVector{3, T}, angular::SVector{3, T})

A twist represents the velocity of a rigid body/frame in 3D space. It is composed of a linear velocity and an angular velocity.

VMRobotControl.Transforms.AxisAngleDerivative — Method

Partial derivative of rotor w.r.t. angle

VMRobotControl.Transforms.angular_velocity_prematrix — Method

angular_velocity_prematrix(r::Rotor)

Return S, which satisfies S*ṙ = ω, where omega is the angular velocity of the frame.

VMRobotControl.Transforms.angular_velocity_prematrix_derivative — Method

angular_velocity_prematrix_derivative(ṙ::SVector)

Return Ṡ, time derivative of angular_velocity_prematrix(r)

VMRobotControl.Transforms.quaternion_derivative_propagation — Method

quaternion_derivative_propagation(r::Rotor)

Returns the matrix E which satisfies E*ω = ṙ, where ω is the angular velocity of rotor r

VMRobotControl.Transforms.quaternion_derivative_propagation_derivative — Method

quaternion_derivative_propagation_derivative(r::Rotor, ω)

Returns the matrix Ė, time derivative of quaternion_derivative_propagation(r)

VMRobotControl.Transforms.quatmul_geodual_bivector_matrix — Method

quatmul_geodual_bivector_matrix(r::Rotor)

Returns a matrix S that corrects from a rotor jacobian ̇r = J⁽ʳ⁾q to an angular velocity jacobian ω = J⁽ʷ⁾̇q, by J⁽ʷ⁾ = S J⁽ʳ⁾.

Equivalent to performing quaternion multiplication with, flipping the direction of the vector and extracting the vector part of the result....

VMRobotControl.Transforms.skew — Method

skew(v::SVector{3})

Return the skew symettric matrix operator performing vector product.

Other API documentation

Here are the remaining docs not included in the main API

VMRobotControl.CompiledMechanism — Type

CompiledMechanism(m::Mechanism)

A compiled version of a Mechanism which is used for efficient kinematics/dynamics computations.

See also: compile, Mechanism

VMRobotControl.CompiledVirtualMechanismSystem — Type

CompiledVirtualMechanismSystem{T, M1, M2, TSC1, TSC2}

A compiled virtual mechanism system is a VirtualMechanismSystem that has been compiled into a form that is efficient for computations/simulation.

VMRobotControl.CompiledCoordID — Type

CompiledCoordID{C}

A unique identifier for a coordinte in a compiled mechanism/virtual mechanism system.

VMRobotControl.reassign_coords — Function

reassign_coords(c, coord_replacement_dict)

For joint/coordinate/component c all references to any coordinate id using the replacement dictionary.

VMRobotControl.GenericComponent — Type

GenericComponent{T} <: ComponentData{T}

Supertype for any components which do not fall cleanly into an inertance, storage or dissipation.

Implement inertance_matrix!, generalized_force!, and stored_energy.

VMRobotControl.Splines.curve_derivative — Function

curve_derivative(t, ::CubicBezierCurve)

First derivative of a cubic bezier curve at t.

VMRobotControl.Splines.curve_second_derivative — Function

curve_second_derivative(t, ::CubicBezierCurve)

Second derivative of a cubic bezier curve at t.

VMRobotControl.ComponentData — Type

ComponentData{T}

Supertype for all mechanical components with eltype T.

VMRobotControl.frame_ancestor_matrix — Function

frame_ancestor_matrix(m::Mechanism)

Returns matrix B such that B[frame_1, frame_2] ⟹ frame_1 is an ancestor of frame 2 in the tree

VMRobotControl.reassign_joints — Function

reassign_joints(c, joint_replacement_dict)

For joint/coordinate/component c replace all references to any joint jointid with joint_replacement_dict[jointid]. Will be recursively applied to fields of the component, such as coordinates.

VMRobotControl.TimeFuncJoint — Type

TimeFuncJoint(f_tf::Function, f_twist::Function, f_vpa::Function)

A TimeFuncJoint represents a joint which has a transform which is a function of time: f_tf(t). The derivatives f_twist and f_vpa must be provided, and checked, by the user.

VMRobotControl._construct_coordinate_graph — Function

_construct_coordinate_graph(coordinates)

Construct a graph of coordinates dependencies on each other.

VMRobotControl.URDF.parseURDFString — Function

parseURDFString(urdfstring::AbstractString, [cfg::URDFParserConfig])

Parse a Mechanism from a URDF string. If cfg is provided, the directory of the URDF file path from the config will be used to find mesh files.

VMRobotControl.RSON.parseRSONString — Function

parseRSONString(rsonstr::AbstractString, [rsonpath::AbstractString], [cfg::RSONParserConfig])

Parse a Mechanism or VirtualMechanismSystem from an RSON string. If rsonpath is not provided, it defaults to ".", which is used for relative paths to locate associated files such as meshes for visuals.

VMRobotControl.RSON.parseRSONVersion — Function

parseRSONVersion(data, cfg)

Will parse the version, and assert that it is compatible with the current parser version. The parsed version will be stored in cfg.rson_version.

The version is expected to be a string of the form "v<major>.<minor>.<patch>", where each part is an integer.

VMRobotControl.AbstractJointData — Type

AbstractJointData{T}

An instance of AbstractJointData represents the data needed to compute the transform between two unspecified frames of reference. When added to a mechanism the frames are specified.

T is the scalar type of the joint (typically Float64), and can be extracted with valuetype(::AbstractJoint).

A joint can be added to a mechanism using the add_joint! function, which requires the joint, the parent frame, and the child frame, creating a MechanismJoint`.

See also

Joint data types: Rigid, Revolute, Prismatic, Rail

Developer Notes

Each joint type must implement the following methods:

_joint_relative_transform (See joint_relative_transform)

_joint_relative_twist (See joint_relative_twist)

_joint_relative_vpa (See joint_relative_vpa)

_joint_relative_acceleration (See joint_relative_acceleration)

_jacobian_columns (See jacobian_column)

VMRobotControl.Spherical — Function

Spherical()
Spherical(tf::Transform)

BROKEN. DOES NOT WORK YET.

Represents a spherical joint between two frames.

See also Revolute

VMRobotControl.Inertance — Type

Inertance{T} <: ComponentData{T}

Supertype for all inertance components, which store kinetic energy and induce inertances and generalized forces in the mechanism.

These components support stored_energy, inertance_matrix!, and generalized_force!, typically by implementing _stored_energy, inertance_matrix!, and _opspace_force.

VMRobotControl.rigid_joint_twist — Function

rigid_joint_twist(tf_parent, tf_child, twist_parent)

Compute the twist of the child frame from the transforms and the twist of the parent frame. Assumes that the frames are rigidly connected.

VMRobotControl._add_jacobian_transpose_times_force! — Function

_jacobian_transpose_times_force!(τ, J, F)

Calculate the product of the transpose of the Jacobian J and the force F, and add it to τ.

VMRobotControl.jacobian_column — Function

jacobian_column(joint, tf_parent, tf_final, q_i)

Returns the jacobian columns Jo and Jw for joint such that if all other joints where stationary, the linear velocity of a frame with transform tf_final would be Jo*q̇ and the angular velocity would be Jw*q̇

VMRobotControl.state_idxs — Function

state_idxs(Σ::CompiledVirtualMechanismSystem)

Returns the indices for getting the state from the ODE vector.

Returns

qʳidxs::UnitRange{Int} - The indices for the robot configuration qᵛidxs::UnitRange{Int} - The indices for the virtual mechanism configuration q̇ʳidxs::UnitRange{Int} - The indices for the robot velocity q̇ᵛidxs::UnitRange{Int} - The indices for the virtual mechanism velocity

VMRobotControl.remake — Function

remake(thing; <keyword arguments>)

Re-construct thing with new field values specified by the keyword arguments. Must have a keyword-constructor.

Adapted from SciMLBase

VMRobotControl.reassign_frames — Function

reassign_frames(c, frame_replacement_dict)

For joint/coordinate/component c replace all references to any frame with frame_replacement_dict[frame]. Will be recursively applied to fields of the component, such as coordinates.

VMRobotControl.MechanismJoint — Type

MechanismJoint(joint, parentFrameID::Int, childFrameID::Int)

Contains jointdata, and the IDs of parent/child frames that the joint connects.

Before a mechanism is compiled, these frame IDs are strings–-the name of the frame–-, after compilation it is an integer, unique for each frame.

See also AbstractJointData, add_joint!

VMRobotControl.Splines.curve_position — Function

curve_position(t, ::CubicBezierCurve)

r(t) is a cubic bezier curve passing through p₁ and p₄ at t=0.0 and 1.0 respectively, with control points p₂ and p₃. It is twice differentiable.

Examples

julia> c = CubicBezierCurve(2.1, 2.3, 2.7, 2.9)
julia> curve_position(0.0, c)
1-element SVector{1, Float64} with indices SOneTo(1):
 2.1

VMRobotControl.EltypeCache — Type

EltypeCache{VALUETYPE, F<:Function}

A cache for storing results of a function with different eltypes. For example, if a Vector{Float64}, and Vector{Float32} cache are both needed, then VALUETYPE can be Vector{T} where T and function can be (VALUETYPE, ELTYPE) -> Vector{ELTYPE}().

VMRobotControl.Dissipation — Type

Dissipation{T} <: ComponentData{T}

Supertype for all dissipative components, which cannot store energy and only dissipate it.

These components support generalized_force!, typically by implementing _opspace_force.